+++ protocols/mdns/src/behaviour/iface/query.rs
@@ -25,7 +25,7 @@ use hickory_proto::{
     rr::{Name, RData},
 };
 use libp2p_core::multiaddr::{Multiaddr, Protocol};
-use libp2p_swarm::address_translation;
+use libp2p_swarm::_address_translation;
 
 use libp2p_identity::PeerId;
 use std::time::Instant;
@@ -178,7 +178,7 @@ impl MdnsResponse {
                 let new_expiration = now + peer.ttl();
 
                 peer.addresses().iter().filter_map(move |address| {
-                    let new_addr = address_translation(address, &observed)?;
+                    let new_addr = _address_translation(address, &observed)?;
                     let new_addr = new_addr.with_p2p(*peer.id()).ok()?;
 
                     Some((*peer.id(), new_addr, new_expiration))