+++ protocols/ping/src/protocol.rs
@@ -19,10 +19,10 @@
 // DEALINGS IN THE SOFTWARE.
 
 use futures::prelude::*;
-use instant::Instant;
 use libp2p_swarm::StreamProtocol;
 use rand::{distributions, prelude::*};
 use std::{io, time::Duration};
+use web_time::Instant;
 
 pub const PROTOCOL_NAME: StreamProtocol = StreamProtocol::new("/ipfs/ping/1.0.0");