+++ swarm/src/connection/pool.rs
@@ -37,7 +37,6 @@ use futures::{
     ready,
     stream::FuturesUnordered,
 };
-use instant::{Duration, Instant};
 use libp2p_core::connection::Endpoint;
 use libp2p_core::muxing::{StreamMuxerBox, StreamMuxerExt};
 use libp2p_core::transport::PortUse;
@@ -52,6 +51,7 @@ use std::{
 };
 use tracing::Instrument;
 use void::Void;
+use web_time::{Duration, Instant};
 
 mod concurrent_dial;
 mod task;