+++ swarm/src/translation.rs
@@ -35,7 +35,8 @@ use libp2p_core::{multiaddr::Protocol, Multiaddr};
 /// address and vice versa.
 ///
 /// If the first [`Protocol`]s are not IP addresses, `None` is returned instead.
-pub fn address_translation(original: &Multiaddr, observed: &Multiaddr) -> Option<Multiaddr> {
+#[doc(hidden)]
+pub fn _address_translation(original: &Multiaddr, observed: &Multiaddr) -> Option<Multiaddr> {
     original.replace(0, move |proto| match proto {
         Protocol::Ip4(_)
         | Protocol::Ip6(_)
@@ -106,7 +107,7 @@ mod tests {
 
         for test in tests.iter() {
             assert_eq!(
-                address_translation(&test.original, &test.observed),
+                _address_translation(&test.original, &test.observed),
                 Some(test.expected.clone())
             );
         }