+++ protocols/identify/CHANGELOG.md
@@ -1,7 +1,8 @@
-## 0.45.0 -- unreleased
+## 0.45.0
 
-- Implement refactored `Transport`.
+- Address translation is moved here from `libp2p-core`.
   See [PR 4568](https://github.com/libp2p/rust-libp2p/pull/4568)
+<!-- Update to libp2p-swarm v0.45.0 -->
 
 ## 0.44.2