+++ protocols/relay/CHANGELOG.md
@@ -1,7 +1,10 @@
-## 0.18.0 -- unreleased
+## 0.18.0
 
-- Implement refactored `Transport`.
-  See [PR 4568](https://github.com/libp2p/rust-libp2p/pull/4568)
+<!-- Update to libp2p-swarm v0.45.0 -->
+
+## 0.17.3
+- Use `web-time` instead of `instant`.
+  See [PR 5347](https://github.com/libp2p/rust-libp2p/pull/5347).
 
 ## 0.17.2