+++ transports/quic/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.11.0 -- unreleased
+## 0.11.0
 
 - Implement refactored `Transport`.
   See [PR 4568](https://github.com/libp2p/rust-libp2p/pull/4568)